Move fonts.ts into metabase/ui
