+++ frontend/src/metabase/admin/embedding/components/ThemeEditor/EditorPanel.tsx
@@ -2,7 +2,6 @@ import { useMemo, useState } from "react";
 import { t } from "ttag";
 
 import type { EmbeddingThemeEditorResult } from "metabase/admin/embedding/hooks/use-embedding-theme-editor";
-import type { MetabaseFontFamily } from "metabase/embedding-sdk/theme/fonts";
 import {
   Box,
   Button,
@@ -17,6 +16,7 @@ import {
   TextInput,
   UnstyledButton,
 } from "metabase/ui";
+import type { MetabaseFontFamily } from "metabase/ui/fonts";
 
 import { ColorSwatchCard } from "./ColorSwatchCard";
 import {
